Revert move of SDKEventToFlow
